backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
